+++ push_notifications/__init__.py
@@ -1,6 +1,6 @@
 __author__ = "Jerome Leclanche"
 __email__ = "jerome@leclan.ch"
-__version__ = "1.5.2"
+__version__ = "1.5.3"
 
 
 class NotificationError(Exception):