+++ push_notifications/migrations/0005_auto_20160909_1035.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('push_notifications', '0004_firefoxdevice'),
+    ]
+
+    operations = [
+        migrations.AlterModelOptions(
+            name='firefoxdevice',
+            options={'verbose_name': 'Firefox device'},
+        ),
+    ]